Migrate components/formulas
